--- langgraph_multi_agent_supervisor/__init__.py
@@ -1,3 +1,0 @@
-from langgraph_multi_agent_supervisor.supervisor import create_supervisor
-
-__all__ = ["create_supervisor"]
