+++ langgraph_supervisor/__init__.py
@@ -1,0 +1,3 @@
+from langgraph_supervisor.supervisor import create_supervisor
+
+__all__ = ["create_supervisor"]
