+++ pyproject.toml
@@ -1,7 +1,10 @@
 [project]
-name = "langgraph-multi-agent-supervisor"
+name = "langgraph-supervisor"
 version = "0.0.1"
 description = "An implementation of a supervisor multi-agent architecture using LangGraph"
+authors = [
+    {name = "Vadym Barda", email = "19161700+vbarda@users.noreply.github.com "}
+]
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
